Post project fb & linkedin 2/Posting/Post_project_fb/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
fastapi==0.109.2
idna==3.10
linkedin-api==2.0.3
marshmallow==3.26.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
python-dotenv==1.0.1
python-facebook-api==0.20.1
requests==2.32.3
requests-oauthlib==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.1